Fixing an issue with the automatic assignment of permissions.
In the last release, when this script was run, it set all files in
the script directory to executable.  This caused the LICENSE.txt
file to be set to executable.  This commit fixes this issue by setting
permissions on any file in this folder with a specified extension
to read/write only.